+++ infra/repository/github_repository.tf
@@ -25,4 +25,18 @@ resource "github_repository" "this" {
 
   archive_on_destroy  = true
   allow_update_branch = true
+
+  security_and_analysis {
+    secret_scanning {
+      status = "enabled"
+    }
+
+    secret_scanning_push_protection {
+      status = "enabled"
+    }
+
+    advanced_security {
+      status = "enabled"
+    }
+  }
 }