+++ infra/repository/locals.tf
@@ -21,7 +21,7 @@ locals {
     secrets = {
       "ARM_CLIENT_ID" = data.azurerm_user_assigned_identity.identity_prod_cd.client_id
     }
-    reviewers_teams = ["io-platform-contributors"]
+    reviewers_teams = ["io-platform-contributors", "engineering-team-cloud-eng"]
   }
 
   app_cd = {