+++ infra/resources/modules/commons/network.tf
@@ -43,6 +43,7 @@ resource "azurerm_private_dns_zone_virtual_network_link" "blob_core_private_vnet
 
   tags = var.tags
 }
+
 resource "azurerm_private_dns_zone_virtual_network_link" "queue_core_private_vnet" {
   name                  = azurerm_virtual_network.vnet.name
   resource_group_name   = azurerm_resource_group.net_rg.name
@@ -52,6 +53,7 @@ resource "azurerm_private_dns_zone_virtual_network_link" "queue_core_private_vne
 
   tags = var.tags
 }
+
 resource "azurerm_private_dns_zone_virtual_network_link" "table_core_private_vnet" {
   name                  = azurerm_virtual_network.vnet.name
   resource_group_name   = azurerm_resource_group.net_rg.name
@@ -78,15 +80,6 @@ resource "azurerm_private_dns_zone_virtual_network_link" "link" {
 # Private endpoints
 #
 
-module "pendpoints_snet" {
-  source               = "github.com/pagopa/terraform-azurerm-v3.git//subnet?ref=v8.7.0"
-  name                 = "${local.project}-pendpoints-snet-01"
-  address_prefixes     = var.snet_pendpoints_address_space
-  resource_group_name  = azurerm_resource_group.net_rg.name
-  virtual_network_name = azurerm_virtual_network.vnet.name
-
-  private_endpoint_network_policies_enabled = false
-}
 
 resource "azurerm_private_endpoint" "sql" {
   name                = format("%s-private-endpoint-sql-01", local.project)
@@ -209,7 +202,7 @@ resource "azurerm_virtual_network_peering" "vnet_to_vnet_common" {
   allow_virtual_network_access = true
   allow_forwarded_traffic      = false
   allow_gateway_transit        = false
-  use_remote_gateways          = false
+  use_remote_gateways          = true
 }
 
 resource "azurerm_virtual_network_peering" "vnet_common_to_vnet" {
@@ -221,6 +214,6 @@ resource "azurerm_virtual_network_peering" "vnet_common_to_vnet" {
 
   allow_virtual_network_access = true
   allow_forwarded_traffic      = false
-  allow_gateway_transit        = false
+  allow_gateway_transit        = true
   use_remote_gateways          = false
 }